Refactor amp and non-amp single post into one component
